requirements.txt
python-decouple
telethon
wheel